# Improve held weight computations