--- a/pol-core/pol/uoscrobj.cpp
+++ b/pol-core/pol/uoscrobj.cpp
@@ -3835,9 +3835,9 @@ Bscript::BObjectImp* UContainer::set_script_member_id_double( const int id, doub
   {
     if ( container )
     {
-      container->add_bulk( 0, -static_cast<int>( weight() ) );
+      int oldweight = weight();
       held_weight_multiplier( value );
-      container->add_bulk( 0, weight() );
+      container->add_bulk( 0, weight() - oldweight );
     }
     else
     {