`held_weight_multiplier` introduced recursive call with exponential complexity which made working with deeply nested containers very slow, this PR optimizes the computations.

Before
```
Loading an item took load=0ms read=0ms store=11554ms (item=1884893733, container=1884936625, depth=28, callcount=268435455)
Loading an item took load=0ms read=0ms store=22974ms (item=1884893734, container=1884893733, depth=29, callcount=536870911)
Loading an item took load=0ms read=0ms store=46285ms (item=1884891786, container=1884893734, depth=30, callcount=1073741823)
Loading an item took load=0ms read=0ms store=92698ms (item=1884891787, container=1884891786, depth=31, callcount=2147483647)
Loading an item took load=0ms read=0ms store=185153ms (item=1884889478, container=1884891787, depth=32, callcount=4294967295)
Loading an item took load=0ms read=0ms store=368158ms (item=1884887796, container=1884889478, depth=33, callcount=8589934591)
Loading an item took load=0ms read=0ms store=722447ms (item=1884889479, container=1884887796, depth=34, callcount=17179869183)
```

After
```
Loading an item took load=0ms read=0ms store=0ms (item=1884893733, container=1884936625, depth=28, callcount=28)
Loading an item took load=0ms read=0ms store=0ms (item=1884893734, container=1884893733, depth=29, callcount=29)
Loading an item took load=0ms read=0ms store=0ms (item=1884891786, container=1884893734, depth=30, callcount=30)
Loading an item took load=0ms read=0ms store=0ms (item=1884891787, container=1884891786, depth=31, callcount=31)
Loading an item took load=0ms read=0ms store=0ms (item=1884889478, container=1884891787, depth=32, callcount=32)
Loading an item took load=0ms read=0ms store=0ms (item=1884887796, container=1884889478, depth=33, callcount=33)
Loading an item took load=0ms read=0ms store=0ms (item=1884889479, container=1884887796, depth=34, callcount=34)
```